% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castDT.R
\name{castDT}
\alias{castDT}
\title{Convert to \code{data.table}}
\usage{
castDT(data, keep, drop, copy = TRUE)
}
\arguments{
\item{data}{Any of the types supported by \code{\link[data.table:as.data.table]{data.table::as.data.table()}} OR
a single character string. If a string, checked against a case-insensitive
pattern \verb{^[^ ]+\\\\.rds$} (i.e. a path with no spaces and ending with ".rds").
If matched, uses \code{\link[=readRDS]{readRDS()}}, otherwise uses \code{\link[data.table:fread]{data.table::fread()}}.}

\item{keep}{columns to select; others dropped. May be: a \code{character()},
an \code{integer()} or a named \code{list()}. If a named list, the values in the list
can provide default values and/or casting operations.
\itemize{
\item if the value is \code{NULL}, no default or casting; i.e. this is a column to
keep, but there are no transformations to apply or defaults to set.
\item if the value satisfies \code{\link[=is.function]{is.function()}}: just a transformation
\item if the value satisfies \code{\link[=is.list]{is.list()}}: both a transformation and default
value
\item otherwise: the default value that will be set if the column is not present
}}

\item{drop}{columns to drop; others selected. May be: a \code{character()} or
\code{integer()}. Emits a warning if dropped columns are not present.}

\item{copy}{Logical; if \code{TRUE} (default), a new \code{data.table} is returned;
if \code{FALSE}, \code{data} will be modified in place if already an R object.}
}
\value{
A \code{data.table}; the returned object will be a copy (default), unless
\code{copy = FALSE}, in which case modifications are made in-place for extant
objects (i.e. if \code{data} is not a character)
}
\description{
Provides consistent casting of \code{data} to \link{data.table}
with error handling.
}
\details{
This function provides a general-purpose tool for basic conversion
operations with \code{data.table}s. It's intended use is as a simplifying,
standardizing interface for data import and column-wise manipulations.

Note: this function will emit any warnings or errors from \code{keep} argument
casting functions.
}
\examples{
mtdt <- castDT(mtcars)
mtdt2 <- castDT(mtdt, keep = c("disp", "hp"))
mtdt3 <- castDT(mtdt, keep = 3:4)
# same as previous
all(mtdt2 == mtdt3)

}
