% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerceDT.R
\name{check_default}
\alias{check_default}
\title{Regularize \code{default} argument}
\usage{
check_default(default, call = parent.frame())
}
\arguments{
\item{default}{columns to select or create with default values. These will
effectively be added to \code{select} if not already present. If they overlap
with \code{select}ed columns, which also specify a transformation, that
transformation will be applied to the default values (if needed).}

\item{call}{the calling environment, used in error messages}
}
\value{
a checked \code{default} list.
}
\description{
Regularize \code{default} argument
}
