% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerceDT.R
\name{check_select}
\alias{check_select}
\title{Regularize \code{select} argument}
\usage{
check_select(select)
}
\arguments{
\item{select}{columns to select; others dropped. May be: a \code{character()},
an \code{integer()} or a named \code{list()}. If a named list, the elements in the list
represent coercion operations for the columns: either NULL (no change),
a string (coerce via \code{as.TYPE}, where that function is present in the global
environment), or a function (coerce via \code{f(s)}).

Emits a warning if selected elements are not present or if any column-wise
coercion would result in \code{NA} values.}
}
\value{
a checked \code{select} value; if \code{select} is a list, will have converted
all the elements to function calls.
}
\description{
Regularize \code{select} argument
}
