% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDT.R
\name{checkDT}
\alias{checkDT}
\title{Check \code{data.table}s}
\usage{
checkDT(data, expect = NULL, forbid = NULL)
}
\arguments{
\item{data}{a \code{data.table}, e.g. as a product of \code{\link[=castDT]{castDT()}}}

\item{expect}{Optional; if \code{NULL} (the default), there are no expected
columns. If a \code{character} vector, \code{castDT} will produce an error indicating
which columns are not present. If a named \code{list}, the names will be required
columns and the \code{list} entries will be used to check the corresponding
columns. If those entries are themselves characters, those will be assumed to
be the class to check to via an \code{is.Class} method; otherwise, they should be
single argument functions that will be used to transform the column.}

\item{forbid}{Optional; if \code{NULL}, ignored. If a character vector,
\code{castDT} will error if any of those columns are present. If anything else,
will error.}
}
\value{
\code{data} itself, assuming passing \code{expect} and \code{forbid}
}
\description{
Provides a checking interface for \code{data.table}s
}
\details{
This function provides a general-purpose tool for common, basic
checking and conversion tasks with \code{data.table}s. It's intended use is as
a simplifying, standardizing interface for raw input checking, not to perform
complex requirement checks or manipulations. It is not, e.g., able to answer
if one-and-only-one of some set of columns are present, or to coerce column
values to a new values based on anything other than their initial value.
}
