% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerceDT-package.R
\docType{package}
\name{coerceDT-package}
\alias{coerceDT-package}
\alias{_PACKAGE}
\title{coerceDT: Ingest-and-check User Input Into Data.table}
\description{
Provides a coerceDT function to ingest-and-check user input into standardized data.tables for use in data science pipelines.
}
\author{
\strong{Maintainer}: Carl Pearson \email{carl.ab.pearson@gmail.com} (\href{https://orcid.org/0000-0003-0701-7860}{ORCID})

Authors:
\itemize{
  \item Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})
}

}
\keyword{internal}
