% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.generalize_keep}
\alias{.generalize_keep}
\title{Regularize \code{keep} argument}
\usage{
.generalize_keep(keep)
}
\arguments{
\item{keep}{columns to select; others dropped. May be: a \code{character()},
an \code{integer()} or a named \code{list()}. If a named list, the values in the list
can provide default values and/or casting operations.
\itemize{
\item if the value is \code{NULL}, no default or casting; i.e. this is a column to
keep, but there are no transformations to apply or defaults to set.
\item if the value satisfies \code{\link[=is.function]{is.function()}}: just a transformation
\item if the value satisfies \code{\link[=is.list]{is.list()}}: both a transformation and default
value
\item otherwise: the default value that will be set if the column is not present
}}
}
\value{
a generalized \code{keep} value: either an \code{integer()}, \code{character()}, or
named \code{list()}
}
\description{
Regularize \code{keep} argument
}
\keyword{internal}
