% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castDT.R
\name{internal_select_drop_convert}
\alias{internal_select_drop_convert}
\title{Select, drop, and convert columns}
\usage{
internal_select_drop_convert(data, select, drop, default)
}
\arguments{
\item{data}{a \code{data.table}}

\item{drop}{columns to drop; others selected. May be: a \code{character()} or
\code{integer()}. Emits a warning if dropped columns are not present.}
}
\description{
Select, drop, and convert columns
}
\details{
ALWAYS modifies \code{data} in place. Does NOT check for consistency of \code{select}
and \code{drop} arguments (i.e. at most one non missing)
}
