% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{loadDT.default}
\alias{loadDT.default}
\alias{loadDT.character}
\alias{loadDT}
\title{Internal Method for Loading Data}
\usage{
\method{loadDT}{default}(data, ..., copy, env = parent.frame())

\method{loadDT}{character}(data, keep, drop, ...)

loadDT(data, keep, drop, copy)
}
\arguments{
\item{data}{Any of the types supported by \code{\link[data.table:as.data.table]{data.table::as.data.table()}} OR
a single character string. If a string, checked against a case-insensitive
pattern \verb{^[^ ]+\\\\.rds$} (i.e. a path with no spaces and ending with ".rds").
If matched, uses \code{\link[=readRDS]{readRDS()}}, otherwise uses \code{\link[data.table:fread]{data.table::fread()}}.}

\item{copy}{Logical; if \code{TRUE} (default), a new \code{data.table} is returned;
if \code{FALSE}, \code{data} will be modified in place if already an R object.}

\item{keep}{columns to select; others dropped. May be: a \code{character()},
an \code{integer()} or a named \code{list()}. If a named list, the values in the list
can provide default values and/or casting operations.
\itemize{
\item if the value is \code{NULL}, no default or casting; i.e. this is a column to
keep, but there are no transformations to apply or defaults to set.
\item if the value satisfies \code{\link[=is.function]{is.function()}}: just a transformation
\item if the value satisfies \code{\link[=is.list]{is.list()}}: both a transformation and default
value
\item otherwise: the default value that will be set if the column is not present
}}

\item{drop}{columns to drop; others selected. May be: a \code{character()} or
\code{integer()}. Emits a warning if dropped columns are not present.}
}
\description{
Internal Method for Loading Data
}
\details{
The \code{keep} argument must have already been \code{\link[=.generalize_keep]{.generalize_keep()}}'d.
}
\keyword{internal}
