% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerceDT.R
\name{coerceDT}
\alias{coerceDT}
\title{Coerce \code{data.table}s}
\usage{
coerceDT(data, select, drop, copy = TRUE)
}
\arguments{
\item{data}{Any of the types supported by \code{\link[data.table:as.data.table]{data.table::as.data.table()}} OR
a single character string. If a string, checked against pattern \verb{\\\\.rds$};
if match, uses \code{\link[=readRDS]{readRDS()}}, otherwise uses \code{\link[data.table:fread]{data.table::fread()}}.}

\item{select}{columns to select; others dropped. May be: a \code{character()},
an \code{integer()} or a named \code{list()}. If a named list, the elements in the list
represent coercion operations for the columns: either NULL (no change),
a string (coerce via \code{as.TYPE}, where that function is present in the global
environment), or a function (coerce via \code{f(s)}).

Emits a warning if selected elements are not present or if any column-wise
coercion would result in \code{NA} values.}

\item{drop}{columns to drop; others selected. May be: a \code{character()} or
\code{integer()}. Emits a warning if dropped columns are not present.}

\item{copy}{Logical; if \code{TRUE} (default), a new \code{data.table} is returned;
if \code{FALSE}, \code{data} will be modified in place if already an R object}
}
\value{
A \code{data.table}; the returned object will be a copy (default), unless
\code{copy = FALSE}, in which case modifications are made in-place for extant
objects (i.e. if \code{data} is not a character)
}
\description{
Provides consistent coercion of \code{data} to \link{data.table}
with error handling.
}
\details{
This function provides a general-purpose tool for basic conversion
operations with \code{data.table}s. It's intended use is as a simplifying,
standardizing interface for data import and column-wise manipulations.
}
\examples{
mtdt <- coerceDT(mtcars)
mtdt2 <- coerceDT(mtdt, select = c("disp", "hp"))
mtdt3 <- coerceDT(mtdt, select = 3:4)
# same as previous
all(mtdt2 == mtdt3)

}
