% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerceDT.R
\name{internal_select_drop_convert}
\alias{internal_select_drop_convert}
\title{Select, drop, and convert columns}
\usage{
internal_select_drop_convert(
  data,
  select,
  drop,
  default,
  call = parent.frame()
)
}
\arguments{
\item{data}{a \code{data.table}}

\item{select}{columns to select; others dropped. May be: a \code{character()},
an \code{integer()} or a named \code{list()}. If a named list, the elements in the list
represent coercion operations for the columns: either NULL (no change),
a string (coerce via \code{as.TYPE}, where that function is present in the global
environment), or a function (coerce via \code{f(s)}).

Emits a warning if selected elements are not present or if any column-wise
coercion would result in \code{NA} values.}

\item{drop}{columns to drop; others selected. May be: a \code{character()} or
\code{integer()}. Emits a warning if dropped columns are not present.}

\item{default}{columns to select or create with default values. These will
effectively be added to \code{select} if not already present. If they overlap
with \code{select}ed columns, which also specify a transformation, that
transformation will be applied to the default values (if needed).}

\item{call}{the calling environment, used in error messages}
}
\description{
Select, drop, and convert columns
}
\details{
ALWAYS modifies \code{data} in place. Does NOT check for consistency of \code{select}
and \code{drop} arguments (i.e. at most one non missing)
}
