% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDT.R
\name{makeDT}
\alias{makeDT}
\title{Make \code{data.table}s}
\usage{
makeDT(data, ..., copy = TRUE)
}
\arguments{
\item{data}{Any of the types supported by \code{\link[data.table:as.data.table]{data.table::as.data.table()}} OR
a single character string. If a string, checked against the pattern \verb{\\\\.rds$};
if match, uses \code{\link[=readRDS]{readRDS()}}, otherwise uses \code{\link[data.table:fread]{data.table::fread()}}.}

\item{copy}{Logical; if \code{TRUE} (default), a new \code{data.table} is returned;
if \code{FALSE}, \code{data} \emph{may} be modified in place, but is not \emph{guaranteed} to be
so. For example, selecting a subset of columns creates a new \code{data.table}
reference.}
}
\value{
A \code{data.table}; the returned object will be a copy (default), unless
\code{copy = FALSE}, in which case modifications \emph{may} be made in-place, though
are not \emph{guaranteed} to have been so
}
\description{
Combines \code{\link[=coerceDT]{coerceDT()}} and \code{\link[=checkDT]{checkDT()}} to provide consistent
coercion of \code{data} to \link{data.table} with error handling.
}
\details{
This function provides a general-purpose tool for common, basic
checking and conversion tasks with \code{data.table}s. It's intended use is as
a simplifying, standardizing interface for raw input checking, not to perform
complex requirement checks or manipulations. It is not, e.g., able to answer
if one-and-only-one of some set of columns are present, or to coerce column
values to a new values based on anything other than their initial value.
}
